Move "UI" into own class
