+++ src/adlpriv.hh
@@ -1,6 +1,9 @@
 #ifndef ADLPRIV_HH
 #define ADLPRIV_HH
 
+#include "../include/adlinput.h"
+#include "../include/adlui.h"
+
 // Setup compiler defines useful for exporting required public API symbols in gme.cpp
 #ifndef ADLMIDI_EXPORT
 #   if defined (_WIN32) && defined(ADLMIDI_BUILD_DLL)
@@ -11,5 +14,8 @@
 #       define ADLMIDI_EXPORT
 #   endif
 #endif
+
+extern ADL_Input Input;
+extern ADL_UserInterface UI;
 
 #endif // ADLPRIV_HH
