+++ src/midiplay.cc
@@ -19,10 +19,10 @@
 
 #include <assert.h>
 
+#include "../include/adlcpp.h"
 #include "../include/adlinput.h"
+#include "../include/adlui.h"
 
-#define SUPPORT_VIDEO_OUTPUT
-#define SUPPORT_PUZZLE_GAME
 
 #ifdef __DJGPP__
 # include <conio.h>
@@ -51,66 +51,29 @@ static const unsigned NewTimerFreq = 209;
 #include "dbopl.h"
 
 #include "adldata.hh"
-
-static const unsigned long PCM_RATE = 48000;
-static const unsigned MaxCards = 100;
-static const unsigned MaxSamplesAtTime = 512; // 512=dbopl limitation
-#else // DJGPP
-static const unsigned MaxCards = 1;
-static const unsigned OPLBase = 0x388;
 #endif
-static unsigned AdlBank    = 0;
-static unsigned NumFourOps = 7;
-static unsigned NumCards   = 2;
-static bool HighTremoloMode   = false;
-static bool HighVibratoMode   = false;
-static bool AdlPercussionMode = false;
-static bool LogarithmicVolumes = false;
-static bool CartoonersVolumes = false;
-static bool QuitFlag = false, FakeDOSshell = false;
-static unsigned SkipForward = 0;
-static bool DoingInstrumentTesting = false;
-static bool QuitWithoutLooping = false;
-static bool WritePCMfile = false;
-static std::string PCMfilepath = "adlmidi.wav";
+
+unsigned AdlBank    = 0;
+unsigned NumFourOps = 7;
+unsigned NumCards   = 2;
+bool HighTremoloMode   = false;
+bool HighVibratoMode   = false;
+bool AdlPercussionMode = false;
+bool LogarithmicVolumes = false;
+bool CartoonersVolumes = false;
+bool QuitFlag = false, FakeDOSshell = false;
+unsigned SkipForward = 0;
+bool DoingInstrumentTesting = false;
+bool QuitWithoutLooping = false;
+bool WritePCMfile = false;
+std::string PCMfilepath = "adlmidi.wav";
 #ifdef SUPPORT_VIDEO_OUTPUT
-static std::string VidFilepath = "adlmidi.mkv";
-static bool WriteVideoFile = false;
+std::string VidFilepath = "adlmidi.mkv";
+bool WriteVideoFile = false;
 #endif
 
-static bool ScaleModulators = false;
-static unsigned WindowLines = 0;
-static bool WritingToTTY;
-
-static unsigned WinHeight()
-{
-    unsigned result =
-        AdlPercussionMode
-        ? std::min(2u, NumCards) * 23
-        : std::min(3u, NumCards) * 18;
-
-    if(WindowLines) result = std::min(WindowLines-1, result);
-    return result;
-}
-
-#if (!defined(__WIN32__) || defined(__CYGWIN__)) && defined(TIOCGWINSZ)
-extern "C" { static void SigWinchHandler(int); }
-static void SigWinchHandler(int)
-{
-    struct winsize w;
-    if(ioctl(2, TIOCGWINSZ, &w) >= 0 || ioctl(1, TIOCGWINSZ, &w) >= 0 || ioctl(0, TIOCGWINSZ, &w) >= 0)
-        WindowLines = w.ws_row;
-}
-#else
-static void SigWinchHandler(int) {}
-#endif
-
-static void GuessInitialWindowHeight()
-{
-    auto s = std::getenv("LINES");
-    if(s && std::atoi(s)) WindowLines = std::atoi(s);
-    SigWinchHandler(0);
-}
+bool ScaleModulators = false;
+bool WritingToTTY;
 
 static const unsigned short Operators[23*2] =
     {// Channels 0-2
@@ -195,23 +158,6 @@ static const adldata& GetAdlIns(unsigned short insno)
 }
 
 
-static const char MIDIsymbols[256+1] =
-"PPPPPPhcckmvmxbd"  // Ins  0-15
-"oooooahoGGGGGGGG"  // Ins 16-31
-"BBBBBBBBVVVVVHHM"  // Ins 32-47
-"SSSSOOOcTTTTTTTT"  // Ins 48-63
-"XXXXTTTFFFFFFFFF"  // Ins 64-79
-"LLLLLLLLpppppppp"  // Ins 80-95
-"XXXXXXXXGGGGGTSS"  // Ins 96-111
-"bbbbMMMcGXXXXXXX"  // Ins 112-127
-"????????????????"  // Prc 0-15
-"????????????????"  // Prc 16-31
-"???DDshMhhhCCCbM"  // Prc 32-47
-"CBDMMDDDMMDDDDDD"  // Prc 48-63
-"DDDDDDDDDDDDDDDD"  // Prc 64-79
-"DD??????????????"  // Prc 80-95
-"????????????????"  // Prc 96-111
-"????????????????"; // Prc 112-127
 
 static const char PercussionMap[256] =
 "\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0"//GM
@@ -236,506 +182,9 @@ static const char PercussionMap[256] =
 
 ADL_Input Input;
 
-#ifdef SUPPORT_PUZZLE_GAME
-#include "puzzlegame.inc"
-#endif
+ADL_UserInterface UI;
 
-#ifdef SUPPORT_VIDEO_OUTPUT
-class UIfontBase {};
-static unsigned UnicodeToASCIIapproximation(unsigned n) { return n; }
-//#include "6x9.inc"
-#include "9x15.inc"
-#endif
 
-class UserInterface
-{
-public:
-    static constexpr unsigned NColumns = 1216/20;
-  #ifdef SUPPORT_VIDEO_OUTPUT
-    static constexpr unsigned VidWidth  = 1216, VidHeight = 2160;
-    static constexpr unsigned FontWidth =   20, FontHeight = 45;
-    static constexpr unsigned TxWidth   = (VidWidth/FontWidth), TxHeight = (VidHeight/FontHeight);
-    unsigned int   PixelBuffer[VidWidth*VidHeight] = {0};
-    unsigned short CharBuffer[TxWidth*TxHeight] = {0};
-    bool           DirtyCells[TxWidth*TxHeight] = {false};
-    unsigned       NDirty = 0;
-  #endif
-  #ifdef __WIN32__
-    void* handle;
-  #endif
-    int x, y, color, txtline, maxy;
-
-    // Text:
-    char background[NColumns][1 + 23*MaxCards];
-    unsigned char backgroundcolor[NColumns][1 + 23*MaxCards];
-    bool          touched[1 + 23*MaxCards]{false};
-    // Notes:
-    char slots[NColumns][1 + 23*MaxCards];
-    unsigned char slotcolors[NColumns][1 + 23*MaxCards];
-
-    bool cursor_visible;
-    char stderr_buffer[256];
-public:
-    UserInterface(): x(0), y(0), color(-1), txtline(0),
-          maxy(0), cursor_visible(true)
-    {
-        GuessInitialWindowHeight();
-      #ifdef __WIN32__
-        handle = GetStdHandle(STD_OUTPUT_HANDLE);
-        GotoXY(41,13);
-        CONSOLE_SCREEN_BUFFER_INFO tmp;
-        GetConsoleScreenBufferInfo(handle,&tmp);
-        if(tmp.dwCursorPosition.X != 41)
-        {
-            // Console is not obeying controls! Probably cygwin xterm.
-            handle = 0;
-        }
-        else
-        {
-            WindowLines = tmp.dwSize.Y;
-            //COORD size = { NColumns, 23*NumCards+5 };
-            //SetConsoleScreenBufferSize(handle,size);
-        }
-      #endif
-      #if (!defined(__WIN32__) || defined(__CYGWIN__)) && defined(TIOCGWINSZ)
-        std::signal(SIGWINCH, SigWinchHandler);
-      #endif
-      #ifdef __DJGPP__
-        color = 7;
-      #endif
-        std::memset(slots, '.',      sizeof(slots));
-        std::memset(background, '.', sizeof(background));
-        std::memset(backgroundcolor, 1, sizeof(backgroundcolor));
-        std::setvbuf(stderr, stderr_buffer, _IOFBF, sizeof(stderr_buffer));
-        RawPrn("\r"); // Ensure cursor is at the x=0 we imagine it being
-        Print(0, 7, true, "Hit Ctrl-C to quit");
-    }
-    void HideCursor()
-    {
-        if(!cursor_visible) return;
-        cursor_visible = false;
-      #ifdef __WIN32__
-        if(handle)
-        {
-          const CONSOLE_CURSOR_INFO info = {100,false};
-          SetConsoleCursorInfo(handle,&info);
-          if(!DoingInstrumentTesting)
-              CheckTetris();
-          return;
-        }
-      #endif
-        if(!DoingInstrumentTesting)
-            CheckTetris();
-#ifdef __DJGPP__
-        { _setcursortype(_NOCURSOR);return;  }
-#endif
-        RawPrn("\33[?25l"); // hide cursor
-    }
-    void ShowCursor()
-    {
-        if(cursor_visible) return;
-        cursor_visible = true;
-        GotoXY(0,maxy); Color(7);
-      #ifdef __WIN32__
-        if(handle)
-        {
-          const CONSOLE_CURSOR_INFO info = {100,true};
-          SetConsoleCursorInfo(handle,&info);
-          return;
-        }
-      #endif
-#ifdef __DJGPP__
-        { _setcursortype(_NORMALCURSOR);return;  }
-#endif
-        RawPrn("\33[?25h"); // show cursor
-        std::fflush(stderr);
-    }
-    void VidPut(char c)
-    {
-    #ifndef SUPPORT_VIDEO_OUTPUT
-        c = c;
-    #else
-        unsigned clr = color, tx = x, ty = y;
-        unsigned cell_index = ty*TxWidth + tx;
-        if(cell_index < TxWidth*TxHeight)
-        {
-            unsigned short what = (clr << 8) + (unsigned(c) & 0xFF);
-            if(what != CharBuffer[cell_index])
-            {
-                CharBuffer[cell_index] = what;
-                if(!DirtyCells[cell_index])
-                {
-                    DirtyCells[cell_index] = true;
-                    ++NDirty;
-                }
-            }
-        }
-    #endif
-    }
-    #ifdef SUPPORT_VIDEO_OUTPUT
-    static unsigned VidTranslateColor(unsigned c)
-    {
-        static const unsigned colors[16] =
-        {
-            0x000000,0x00005F,0x00AA00,0x5F5FAF,
-            0xAA0000,0xAA00AA,0x87875F,0xAAAAAA,
-            0x005F87,0x5555FF,0x55FF55,0x55FFFF,
-            0xFF5555,0xFF55FF,0xFFFF55,0xFFFFFF
-        };
-        return colors[c % 16];
-    }
-    void VidRenderCh(unsigned x,unsigned y, unsigned attr, const unsigned char* bitmap)
-    {
-        unsigned bg = VidTranslateColor(attr >> 4), fg = VidTranslateColor(attr);
-        for(unsigned line=0; line<FontHeight; ++line)
-        {
-            unsigned int* pix = PixelBuffer + (y*FontHeight+line)*VidWidth + x*FontWidth;
-
-            unsigned char bm = bitmap[line*15/FontHeight];
-            for(unsigned w=0; w<FontWidth; ++w)
-            {
-                int shift0 = 7 - w    *9/FontWidth;
-                int shift1 = 7 - (w-1)*9/FontWidth;
-                int shift2 = 7 - (w+1)*9/FontWidth;
-                bool flag = (shift0 >= 0 && (bm & (1u << shift0)));
-                if(!flag && (attr & 8))
-                {
-                    flag = (shift1 >= 0 && (bm & (1u << shift1)))
-                        || (shift2 >= 0 && (bm & (1u << shift2)));
-                }
-                pix[w] = flag ? fg : bg;
-            }
-        }
-    }
-    void VidRender()
-    {
-        static const font9x15 font;
-        if(NDirty)
-        {
-            #pragma omp parallel for schedule(static)
-            for(unsigned scan=0; scan<TxWidth*TxHeight; ++scan)
-                if(DirtyCells[scan])
-                {
-                    --NDirty;
-                    DirtyCells[scan] = false;
-                    VidRenderCh(scan%TxWidth, scan/TxWidth,
-                                CharBuffer[scan] >> 8,
-                                font.GetBitmap() + 15 * font.GetIndex(CharBuffer[scan] & 0xFF));
-                }
-        }
-    }
-    #endif
-    void PutC(char c)
-    {
-      #ifdef __WIN32__
-        if(handle) WriteConsole(handle,&c,1, 0,0);
-        else
-      #endif
-        {
-      #ifdef __DJGPP__
-          putch(c);
-      #else
-          std::fputc(c, stderr);
-      #endif
-        }
-        VidPut(c);
-        ++x; // One letter drawn. Update cursor position.
-    }
-    #ifdef __DJGPP__
-    #define RawPrn cprintf
-    #else
-    static void RawPrn(const char* fmt, ...) __attribute__((format(printf,1,2)))
-    {
-        // Note: should essentially match PutC, except without updates to x
-        va_list ap;
-        va_start(ap, fmt);
-        vfprintf(stderr, fmt, ap);
-        va_end(ap);
-    }
-    #endif
-    int Print(unsigned beginx, unsigned color, bool ln, const char* fmt, va_list ap)
-    {
-        char Line[1024];
-      #ifndef __CYGWIN__
-        int nchars = vsnprintf(Line, sizeof(Line), fmt, ap);
-      #else
-        int nchars = std::vsprintf(Line, fmt, ap); /* SECURITY: POSSIBLE BUFFER OVERFLOW (Cygwin) */
-      #endif
-        //if(nchars == 0) return nchars;
-
-        HideCursor();
-        for(unsigned tx = beginx; tx < NColumns; ++tx)
-        {
-            int n = tx-beginx; // Index into Line[]
-            if(n < nchars && Line[n] != '\n')
-            {
-                GotoXY(tx, txtline);
-                Color( backgroundcolor[tx][txtline] = (/*Line[n] == '.' ? 1 :*/ color) );
-                PutC( background[tx][txtline] = Line[n] );
-            }
-            else //if(background[tx][txtline]!='.' && slots[tx][txtline]=='.')
-            {
-                if(!ln) break;
-                GotoXY(tx,txtline);
-                Color( backgroundcolor[tx][txtline] = 1);
-                PutC( background[tx][txtline] = '.' );
-            }
-        }
-        std::fflush(stderr);
-
-        if(ln)
-        {
-            txtline = (txtline + 1) % WinHeight();
-        }
-
-        return nchars;
-    }
-    int Print(unsigned beginx, unsigned color, bool ln, const char* fmt, ...) __attribute__((format(printf,5,6)))
-    {
-        va_list ap;
-        va_start(ap, fmt);
-        int r = Print(beginx, color, ln, fmt, ap);
-        va_end(ap);
-        return r;
-    }
-    int PrintLn(const char* fmt, ...) __attribute__((format(printf,2,3)))
-    {
-        va_list ap;
-        va_start(ap, fmt);
-        int r = Print(2/*column*/, 8/*color*/, true/*line*/, fmt, ap);
-        va_end(ap);
-        return r;
-    }
-    void IllustrateNote(int adlchn, int note, int ins, int pressure, double bend)
-    {
-        HideCursor();
-    #if 1
-        int notex = 2 + (note+55)%(NColumns-3);
-        int limit = WinHeight(), minline = 3;
-
-        int notey = minline + adlchn;
-        notey %= std::max(1, limit-minline);
-        notey += minline;
-        notey %= limit;
-
-        char illustrate_char = background[notex][notey];
-        if(pressure > 0)
-        {
-            illustrate_char = MIDIsymbols[ins];
-            if(bend < 0) illustrate_char = '<';
-            if(bend > 0) illustrate_char = '>';
-        }
-        else if(pressure < 0)
-        {
-            illustrate_char = '%';
-        }
-        // Special exceptions for '.' (background slot)
-        //                        '&' (tetris edges)
-        Draw(notex,notey,
-            pressure!=0
-                ? AllocateColor(ins)        /* use note's color if active */
-                : illustrate_char=='.' ? backgroundcolor[notex][notey]
-                : illustrate_char=='&' ? 1
-                : 8,
-            illustrate_char);
-    #endif
-        std::fflush(stderr);
-    }
-
-    void Draw(int notex,int notey, int color, char ch)
-    {
-        if(slots[notex][notey] != ch
-        || slotcolors[notex][notey] != color)
-        {
-            slots[notex][notey] = ch;
-            slotcolors[notex][notey] = color;
-            GotoXY(notex, notey);
-            Color(color);
-            PutC(ch);
-
-            if(!touched[notey])
-            {
-                touched[notey] = true;
-
-                GotoXY(0, notey);
-                for(int tx=0; tx<int(NColumns); ++tx)
-                {
-                    if(slots[tx][notey] != '.')
-                    {
-                        Color(slotcolors[tx][notey]);
-                        PutC(slots[tx][notey]);
-                    }
-                    else
-                    {
-                        Color(backgroundcolor[tx][notey]);
-                        PutC(background[tx][notey]);
-                    }
-                }
-            }
-        }
-    }
-
-    void IllustrateVolumes(double left, double right)
-    {
-        const unsigned maxy = WinHeight();
-        const unsigned white_threshold  = maxy/23;
-        const unsigned red_threshold    = maxy*4/23;
-        const unsigned yellow_threshold = maxy*8/23;
-
-        double amp[2] = {left*maxy, right*maxy};
-        for(unsigned y=2; y<maxy; ++y)
-            for(unsigned w=0; w<2; ++w)
-            {
-                char c = amp[w] > (maxy-1)-y ? '|' : background[w][y+1];
-                Draw(w,y+1,
-                     c=='|' ? y<white_threshold ? 15
-                            : y<red_threshold ? 12
-                            : y<yellow_threshold ? 14
-                            : 10 : (c=='.' ? 1 : 8),
-                     c);
-            }
-        std::fflush(stderr);
-    }
-
-    // Move tty cursor to the indicated position.
-    // Movements will be done in relative terms
-    // to the current cursor position only.
-    void GotoXY(int newx, int newy)
-    {
-        // Record the maximum line count seen
-        if(newy > maxy) maxy = newy;
-        // Go down with '\n' (resets cursor at beginning of line)
-        while(newy > y)
-        {
-            std::fputc('\n', stderr); y+=1; x=0;
-        }
-      #ifdef __WIN32__
-        if(handle)
-        {
-          CONSOLE_SCREEN_BUFFER_INFO tmp;
-          GetConsoleScreenBufferInfo(handle, &tmp);
-          COORD tmp2 = { x = newx, tmp.dwCursorPosition.Y } ;
-          if(newy < y) { tmp2.Y -= (y-newy); y = newy; }
-          SetConsoleCursorPosition(handle, tmp2);
-        }
-      #endif
-#ifdef __DJGPP__
-        { gotoxy(x=newx, wherey()-(y-newy)); y=newy; return; }
-#endif
-        // Go up with \33[A
-        if(newy < y) { RawPrn("\33[%dA", y-newy); y = newy; }
-        // Adjust X coordinate
-        if(newx != x)
-        {
-            // Use '\r' to go to column 0
-            if(newx == 0) // || (newx<10 && std::abs(newx-x)>=10))
-                { std::fputc('\r', stderr); x = 0; }
-            // Go left  with \33[D
-            if(newx < x) RawPrn("\33[%dD", x-newx);
-            // Go right with \33[C
-            if(newx > x) RawPrn("\33[%dC", newx-x);
-            x = newx;
-        }
-    }
-    // Set color (4-bit). Bits: 1=blue, 2=green, 4=red, 8=+intensity
-    void Color(int newcolor)
-    {
-        if(color != newcolor)
-        {
-          #ifdef __WIN32__
-            if(handle)
-              SetConsoleTextAttribute(handle, newcolor);
-            else
-          #endif
-#ifdef __DJGPP__
-            textattr(newcolor);
-            if(0)
-#endif
-            {
-              static const char map[8+1] = "04261537";
-              RawPrn("\33[0;%s40;3%c", (newcolor&8) ? "1;" : "", map[newcolor&7]);
-              // If xterm-256color is used, try using improved colors:
-              //        Translate 8 (dark gray) into #003366 (bluish dark cyan)
-              //        Translate 1 (dark blue) into #000033 (darker blue)
-              if(newcolor==8) RawPrn(";38;5;24;25");
-              if(newcolor==6) RawPrn(";38;5;101;25");
-              if(newcolor==3) RawPrn(";38;5;61;25");
-              if(newcolor==1) RawPrn(";38;5;17;25");
-              RawPrn("m");
-            }
-            color = newcolor;
-        }
-    }
-    // Choose a permanent color for given instrument
-    int AllocateColor(int ins)
-    {
-        static char ins_colors[256] = { 0 }, ins_color_counter = 0;
-        if(ins_colors[ins])
-            return ins_colors[ins];
-        if(ins & 0x80)
-        {
-            static const char shuffle[] = {2,3,4,5,6,7};
-            return ins_colors[ins] = shuffle[ins_color_counter++ % 6];
-        }
-        else
-        {
-            static const char shuffle[] = {10,11,12,13,14,15};
-            return ins_colors[ins] = shuffle[ins_color_counter++ % 6];
-        }
-    }
-
-    bool DoCheckTetris()
-    {
-    #ifdef SUPPORT_PUZZLE_GAME
-        static ADLMIDI_PuzzleGame::TetrisAI    player(2);
-        static ADLMIDI_PuzzleGame::TetrisHuman computer(31);
-
-        int a = player.GameLoop();
-        int b = computer.GameLoop();
-
-        if(a >= 0 && b >= 0)
-        {
-            player.incoming   += b;
-            computer.incoming += a;
-        }
-        return player.DelayOpinion() >= 0
-            && computer.DelayOpinion() >= 0;
-    #else
-        return true;
-    #endif
-    }
-
-    bool TetrisLaunched = false;
-    bool CheckTetris()
-    {
-        if(TetrisLaunched) return DoCheckTetris();
-        return true;
-    }
-} UI;
-
-#ifdef SUPPORT_PUZZLE_GAME
-namespace ADLMIDI_PuzzleGame
-{
-    static void PutCell(int x, int y, unsigned cell)
-    {
-        static const unsigned char valid_attrs[] = {8,6,5,3};
-        unsigned char ch = cell, attr = cell >> 8;
-        int height = WinHeight();//std::min(NumCards*18, 50u);
-        y = std::max(0, int(std::min(height, 40) - 25 + y));
-        if(ch == 0xDB) ch = '#';
-        if(ch == 0xB0) ch = '*';
-        if(attr != 1) attr = valid_attrs[attr % sizeof(valid_attrs)];
-
-        //bool diff = UI.background[x][y] != UI.slots[x][y];
-        UI.backgroundcolor[x][y] = attr;
-        UI.background[x][y]      = ch;
-        UI.GotoXY(x,y);
-        UI.Color(attr);
-        UI.PutC(ch);
-        //UI.Draw(x,y, attr, ch);
-    }
-}
-#endif
 
 struct OPL3
 {
